move getting and updating service arguments to snap package
